+++ NEWS.md
@@ -10,6 +10,10 @@
 
 - Added the `keep_md` argument to `blogdown::html_page()` (thanks, @lazappi, #445).
 
+## MINOR CHANGES
+
+- Images `tn.png` and `screenshot.png` under the `images/` directory of a theme will be deleted in `blogdown::install_theme()` because these are screnshots of a theme and don't affect the theme's function.
+
 # CHANGES IN blogdown VERSION 0.20
 
 ## BUG FIXES